requirements.txt
line-bot-sdk
flask
gunicorn
liffpy
pymongo
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
pytz
openai